Refactor: Move _extract_user_message_text to utils module and apply classification improvements
