+++ utils/__init__.py
@@ -1,0 +1,1 @@
+# Utility functions for the tri-tier private AI system
